app/pyproject.toml
[tool.poetry]
name = "messenger"
version = "0.1.0"
description = ""
authors = ["pavelhardzei <pavelgordei11@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pre-commit = "^2.16.0"
Django = "^3.2.9"
psycopg2-binary = "^2.9.2"
gunicorn = "^20.1.0"
djangorestframework = "^3.12.4"
channels = "^3.0.4"
channels-redis = "^3.3.1"
uritemplate = "^4.1.1"
PyYAML = "^6.0"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.1.0"
pytest-asyncio = "^0.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
